requirements.txt
asgiref==3.7.2
Django==5.0.1
django-etc==1.4.0
django-modeltranslation==0.18.11
environs==10.3.0
gunicorn==21.2.0
marshmallow==3.20.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
segno==1.6.0
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
